remove warning casting from double to int

+++ src/tlssh.cc
@@ -235,8 +235,8 @@ mainloop(FDWrap &terminal)
 
                 int timeout = -1;
                 if (options.keepalive != 0) {
-                        timeout = 1000 * (options.keepalive
-                                          - (now - last_keepalive_sent));
+                        timeout = (int)(1000* (options.keepalive
+                                               - (now - last_keepalive_sent)));
                         // protect against rounding errors
                         timeout = std::max(timeout, 0);
                 }
